# docs/openapi.yaml
openapi: 3.0.3
info:
    title: Goveto Edge Control API
    version: 0.1.0
    description: Control-plane API for managing edge clusters, nodes, sites, certificates, publish, purge and analytics.
servers:
    - url: /
...
paths:
...
    /api/v1/clusters/{cluster_id}/analytics/summary:
        get:
            operationId: summary
            summary: Analytics summary
            description: Aggregate request/traffic summary for a time range (from/to).
            tags:
                - analytics
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: from
                  in: query
                  schema:
                    type: string
                    format: date-time
                    example: "2026-01-02T15:04:05Z"
                - name: site_id
                  in: query
                  schema:
                    type: string
                - name: to
                  in: query
                  schema:
                    type: string
                    format: date-time
                    example: "2026-01-02T15:04:05Z"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
...
    /api/v1/clusters/{cluster_id}/nodes:
        get:
            operationId: list
            summary: List nodes
            description: List nodes in the cluster with addresses and site config versions.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        post:
            operationId: create
            summary: Create node
            description: Create a node and enqueue remote installation with SSH credentials.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                addresses:
                                    type: array
                                    items:
                                        type: string
                                dns_line_ids:
                                    type: array
                                    items:
                                        type: string
                                group_id:
                                    type: string
                                name:
                                    type: string
                                region_id:
                                    type: string
                                ssh:
                                    $ref: '#/components/schemas/node_SSHInstallInput'
            responses:
                "202":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        properties:
                                            id:
                                                type: object
                                            status:
                                                type: string
                                                enum:
                                                    - PENDING
                                        required:
                                            - id
                                            - status
    /api/v1/clusters/{cluster_id}/nodes/{node_id}:
        get:
            operationId: get
            summary: Get node
            description: Get a single node with addresses, site config versions and cache config.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        delete:
            operationId: deleteNode
            summary: Delete node
            description: Delete a node and related credentials, addresses and config records.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
    /api/v1/clusters/{cluster_id}/nodes/{node_id}/addresses:
        post:
            operationId: addAddress
            summary: Add node address
            description: Add an IP address to a node; optionally mark it as primary.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                address:
                                    type: string
                                primary:
                                    type: boolean
            responses:
                "201":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
    /api/v1/clusters/{cluster_id}/nodes/{node_id}/cache-config:
        get:
            operationId: getCacheConfig
            summary: Get node cache config
            description: Get disk cache configuration for a node.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        put:
            operationId: updateCacheConfig
            summary: Update node cache config
            description: Update node disk cache settings and attempt to sync to the edge agent.
            tags:
                - nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                auto_max_size:
                                    type: boolean
                                cache_directory:
                                    type: string
                                max_disk_usage_percent:
                                    type: number
                                max_size_bytes:
                                    type: number
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        properties:
                                            cache_config:
                                                type: object
                                            synced:
                                                type: boolean
                                                enum:
                                                    - false
                                        required:
                                            - cache_config
                                            - synced
    /api/v1/clusters/{cluster_id}/nodes/{node_id}/disable:
        post:
            operationId: disableNode
            summary: post /api/v1/clusters/{cluster_id}/nodes/{node_id}/disable
            tags:
                - clusters-nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        properties:
                                            id:
                                                type: string
                                            status:
                                                type: string
                                                enum:
                                                    - DISABLED
                                        required:
                                            - id
                                            - status
    /api/v1/clusters/{cluster_id}/nodes/{node_id}/dns-lines:
        put:
            operationId: updateDNSLines
            summary: put /api/v1/clusters/{cluster_id}/nodes/{node_id}/dns-lines
            tags:
                - clusters-nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                dns_line_ids:
                                    type: array
                                    items:
                                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        properties:
                                            dns_line_ids:
                                                type: array
                                                items:
                                                    type: string
                                            node_id:
                                                type: string
                                        required:
                                            - dns_line_ids
                                            - node_id
    /api/v1/clusters/{cluster_id}/nodes/{node_id}/enable:
        post:
            operationId: enableNode
            summary: post /api/v1/clusters/{cluster_id}/nodes/{node_id}/enable
            tags:
                - clusters-nodes
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: node_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "202":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        properties:
                                            id:
                                                type: string
                                            message:
                                                type: string
                                                enum:
                                                    - waiting for health check
                                            status:
                                                type: string
                                                enum:
                                                    - OFFLINE
                                        required:
                                            - id
                                            - message
                                            - status
...
    /api/v1/clusters/{cluster_id}/sites/{site_id}/publish:
        get:
            operationId: listJobs
            summary: List publish jobs
            description: List recent publish jobs for a site.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        post:
            operationId: enqueue
            summary: Enqueue publish
            description: Enqueue a site configuration publish job.
            tags:
                - publish
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "202":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
...
    /api/v1/clusters/{cluster_id}/sites/{site_id}/purge:
        get:
            operationId: list
            summary: List purge jobs
            description: List recent cache purge jobs for a site.
            tags:
                - purge
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
        post:
            operationId: enqueue
            summary: Enqueue purge
            description: Enqueue a cache purge job for the site.
            tags:
                - purge
            security:
                - cookie_session: []
            x-middlewares:
                - Session
                - RequireAuth
                - Require
            parameters:
                - name: cluster_id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: site_id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                type:
                                    $ref: '#/components/schemas/model_PurgeType'
                                value:
                                    type: string
            responses:
                "202":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        enum:
                                            - ok
                                    data:
                                        type: object
                                        additionalProperties:
                                            type: object
...
components:
    schemas:
...
        model_PurgeType:
            type: string
...
        node_SSHInstallInput:
            type: object
            properties:
                entry_ip:
                    type: string
                passphrase:
                    type: string
                password:
                    type: string
                port:
                    type: number
                private_key:
                    type: string
                user:
                    type: string
...
    securitySchemes:
        cookie_session:
            type: apiKey
            in: cookie
            name: session